+++ tests/acceptance_tests/test_app.py
@@ -587,7 +587,8 @@ SELECT ?link ?name ?format ?description ?kind {{
                         "args": {
                             "query": "SELECT * { ?substation a cim:Substation }"
                         },
-                        "errorOutput": "The following prefixes are undefined: cim"
+                        "errorOutput": "The following prefixes are undefined: cim",
+                        "graphdbRepositoryId": "cim"
                     }
                 ]
             },
@@ -794,7 +795,8 @@ SELECT ?link ?name ?format ?description ?kind {{
                             "query": "SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> }"
                         },
                         "query": "SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> }",
-                        "queryType": "sparql"
+                        "queryType": "sparql",
+                        "graphdbRepositoryId": "cim"
                     }
                 ]
             },